float work correctly

+++ filterCoef.h
@@ -4,6 +4,6 @@
 #include "stdint.h"
 
 #define filterOrder 101
-extern int16_t filterCoef[filterOrder];
+extern float filterCoef[filterOrder];
 
 #endif // FILTERCOEF_H_
